fix: rename wrong file name
