+++ libs/coin-modules/coin-stacks/src/constants.ts
